refactor: remove old ControlsDropdownOld and adjust new ControlsDropdown to handle all of its work
